+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### WinchControllers v0.6 2025-05-28
+### WinchControllers v0.5.1 2025-05-28
 #### Added
 - all public functions document
 - add page `Performance Indicators` to documentation
@@ -10,5 +10,5 @@
 #### Changed
 - reduced `df_high` in `wc_settings.yaml` to reduce oscillations
 - all examples are making use of `wc_settings.yaml` now
-- the function `get_v_error` returns now `NaN` instead of zero when the speed controller is inactive
+- the function `get_v_err` returns now `NaN` instead of zero when the speed controller is inactive
 - improved the plots of the examples